Fix a crash when cancelling the OFX update dialog while waiting for operations that occur until the TransferJob is created. These operations can be things like waiting for KWallet, waiting for libofx_request_statement, etc.
BUG: 281728

+++ kmymoney/plugins/ofximport/dialogs/kofxdirectconnectdlg.cpp
@@ -219,7 +219,8 @@ void KOfxDirectConnectDlg::slotOfxFinished(KJob* /* e */)
 
 void KOfxDirectConnectDlg::reject(void)
 {
-  m_job->kill();
+  if (m_job)
+    m_job->kill();
   if (m_tmpfile) {
     m_tmpfile->close();
     delete m_tmpfile;
